+++ recipes/default/phpunit/run.sh
@@ -11,7 +11,7 @@ test -d $ROOT_PATH/source || (mkdir -p $ROOT_PATH/source && cp -r $SCRIPT_PATH/e
 
 make up
 
-docker-compose exec -T php composer update --no-interaction
-docker-compose exec -T php composer require phpunit/phpunit --no-interaction
+docker compose exec -T php composer update --no-interaction
+docker compose exec -T php composer require phpunit/phpunit --no-interaction
 
 echo "Done!"